+++ traffic_ops/traffic_ops_golang/deliveryservice/letsencrypt_dns_challenge.go
@@ -22,9 +22,9 @@ package deliveryservice
 import (
 	"database/sql"
 	"errors"
-	"fmt"
-	"github.com/apache/trafficcontrol/traffic_ops/traffic_ops_golang/api"
 	"net/http"
+
+	"github.com/apache/trafficcontrol/traffic_ops/traffic_ops_golang/api"
 )
 
 type DnsRecord struct {
@@ -32,24 +32,7 @@ type DnsRecord struct {
 	Record *string `json:"record" db:"record"`
 }
 
-func GetDnsChallengeRecord(w http.ResponseWriter, r *http.Request) {
-	inf, userErr, sysErr, errCode := api.NewInfo(r, []string{"fqdn"}, nil)
-	if userErr != nil || sysErr != nil {
-		api.HandleErr(w, r, inf.Tx.Tx, errCode, userErr, sysErr)
-		return
-	}
-	defer inf.Close()
-
-	dnsRecord, err := getDnsRecord(inf.Tx.Tx, inf.Params["fqdn"])
-	if err != nil {
-		api.HandleErr(w, r, inf.Tx.Tx, http.StatusInternalServerError, nil, errors.New("checking dns record: "+err.Error()))
-		return
-	}
-
-	api.WriteResp(w, r, dnsRecord)
-}
-
-func GetAllDnsChallengeRecords(w http.ResponseWriter, r *http.Request) {
+func GetDnsChallengeRecords(w http.ResponseWriter, r *http.Request) {
 	inf, userErr, sysErr, errCode := api.NewInfo(r, nil, nil)
 	if userErr != nil || sysErr != nil {
 		api.HandleErr(w, r, inf.Tx.Tx, errCode, userErr, sysErr)
@@ -57,29 +40,23 @@ func GetAllDnsChallengeRecords(w http.ResponseWriter, r *http.Request) {
 	}
 	defer inf.Close()
 
-	dnsRecord, err := getDnsRecords(inf.Tx.Tx)
+	getQuery := `SELECT fqdn, record FROM dnschallenges`
+
+	if inf.Params["fqdn"] != "" {
+		getQuery += ` where fqdn = '` + inf.Params["fqdn"] + `'`
+	}
+
+	dnsRecord, err := getDnsRecords(inf.Tx.Tx, getQuery)
 	if err != nil {
 		api.HandleErr(w, r, inf.Tx.Tx, http.StatusInternalServerError, nil, errors.New("checking dns records: "+err.Error()))
 		return
 	}
-
 	api.WriteResp(w, r, dnsRecord)
 }
 
-func getDnsRecord(tx *sql.Tx, fqdn string) (DnsRecord, error) {
-	record := DnsRecord{}
-	if err := tx.QueryRow(`SELECT fqdn, record FROM dnschallenges where fqdn = $1`, fqdn).Scan(&record.Fqdn, &record.Record); err != nil {
-		if err == sql.ErrNoRows {
-			return DnsRecord{}, nil
-		}
-		return DnsRecord{}, fmt.Errorf("querying dns challenge record for fqdn '%v': %v", fqdn, err)
-	}
-	return record, nil
-}
-
-func getDnsRecords(tx *sql.Tx) ([]DnsRecord, error) {
+func getDnsRecords(tx *sql.Tx, getQuery string) ([]DnsRecord, error) {
 	records := []DnsRecord{}
-	rows, err := tx.Query(`SELECT fqdn, record FROM dnschallenges`)
+	rows, err := tx.Query(getQuery)
 	if err != nil {
 		return nil, errors.New("getting dns challenge records: " + err.Error())
 	}