+++ traffic_ops/traffic_ops_golang/routing/routes.go
@@ -981,8 +981,7 @@ func Routes(d ServerData) ([]Route, []RawRoute, http.Handler, error) {
 
 		//Delivery service LetsEncrypt
 		{api.Version{1, 4}, http.MethodPost, `deliveryservices/sslkeys/generate/letsencrypt/?(\.json)?$`, deliveryservice.GenerateLetsEncryptCertificates, auth.PrivLevelReadOnly, Authenticated, nil, 753439052, noPerlBypass},
-		{api.Version{1, 4}, http.MethodGet, `letsencrypt/dnsrecord/?(\.json)?$`, deliveryservice.GetDnsChallengeRecord, auth.PrivLevelReadOnly, Authenticated, nil, 753439054, noPerlBypass},
-		{api.Version{1, 4}, http.MethodGet, `letsencrypt/dnsrecords/?(\.json)?$`, deliveryservice.GetAllDnsChallengeRecords, auth.PrivLevelReadOnly, Authenticated, nil, 753439055, noPerlBypass},
+		{api.Version{1, 4}, http.MethodGet, `letsencrypt/dnsrecords/?(\.json)?$`, deliveryservice.GetDnsChallengeRecords, auth.PrivLevelReadOnly, Authenticated, nil, 753439055, noPerlBypass},
 		{api.Version{1, 4}, http.MethodPost, `letsencrypt/autorenew/?(\.json)?$`, deliveryservice.RenewCertificates, auth.PrivLevelReadOnly, Authenticated, nil, 753439056, noPerlBypass},
 
 		{api.Version{1, 1}, http.MethodGet, `deliveryservices/{id}/health/?(\.json)?$`, deliveryservice.GetHealth, auth.PrivLevelReadOnly, Authenticated, nil, 2034590101, perlBypass},