+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Added an optimistic quorum feature to Traffic Monitor to prevent false negative states from propagating to downstream components in the event of network isolation.
 - Added the ability to fetch users by role
 - Added an API 1.4 endpoint to generate delivery service certificates using Let's Encrypt
-- Added API 1.4 endpoints to GET a single or all records for Let's Encrypt DNS challenge
+- Added an API 1.4 endpoint to GET a single or all records for Let's Encrypt DNS challenge
 - Added an API 1.4 endpoint to renew certificates
 - Traffic Ops Golang Endpoints
   - /api/2.0 for all of the most recent route versions
@@ -22,7 +22,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - /api/1.1/cachegroupparameters/ `(GET, POST)`
   - /api/1.4/deliveryservices/sslkeys/generate/letsencrypt `POST`
   - /api/1.4/letsencrypt/autorenew `POST`
-  - /api/1.4/letsencrypt/dnsrecord `GET`
   - /api/1.4/letsencrypt/dnsrecords `GET`
 
 ### Changed